gained a little bit of performance checking wether sprites should be drawn or destroyed

+++ common/include/Frame.h
@@ -11,6 +11,6 @@ typedef enum {
 	FRAME_32x32
 } FrameSize;
 
-void DrawFrame(FrameSize size, int idx, UINT16 x, UINT16 y, UINT8 flags);
+void DrawFrame(FrameSize size, int idx, UINT8 x, UINT8 y, UINT8 flags);
 
 #endif
